Fix "Playback position not available" error
When initially running player-mpris-tail.py, it displays the current
track being played without issue. However when the track is changed
(whether that's skipping to the next, or restarting current) an
exception is thrown causing the script's output to be replaced with
"ERROR:dbus.connection:Exception in handler for D-Bus signal:".

This catches that exception and sets a somewhat sane default for
speed_us when the track's position isn't available.