Fix missing required positional arguments
When changing track, the dbus library was throwing an error saying the 2
required positional arguments `properties` and `signature` were missing.

Apparently not all events will include those, but as we're only using
the PropertiesChanged signal, we can set it explicitly when we add the
signal receiver. Because it now listens only for the PropertiesChanged
signal, we can safely assume both keyword arguments will be passed.