player-cmus: handle no file information
When CMUS is started for the first time, cmus-remote doesn't return
any track data, which means $duration is an empty string. This causes
the script to fail with an error on line 12 when it tries to evaluate
$duration as an integer.

The fix is to only print any information if we have either a title or
an artist, which means we probably also have a duration as well.